requirements.txt
discord.py[voice]
bs4
Pillow
asyncpg
matplotlib
imgkit
aiohttp
dateparser
pandas
numpy
tabulate
yaml
xmltodict
requests